javadoc comments

+++ src/main/java/pl/polsl/library/model/Role.java
@@ -8,25 +8,49 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
 
+/**
+ * The {@code Role} entity represents a user role and implements the Spring Security GrantedAuthority interface.
+ */
 @Entity
 @Getter
 @Setter
 public class Role implements GrantedAuthority {
+
+    /**
+     * The unique identifier for the role.
+     */
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
 
+    /**
+     * The authority (name) of the role.
+     */
     private String authority;
 
-    public Role(){
+    /**
+     * Default constructor for the Role entity.
+     */
+    public Role() {
         super();
     }
 
-    public Role(String authority){
+    /**
+     * Constructor for creating a role with the specified authority.
+     *
+     * @param authority The authority (name) of the role.
+     */
+    public Role(String authority) {
         this.authority = authority;
     }
 
-    public Role(Integer roleId, String authority){
+    /**
+     * Constructor for creating a role with the specified ID and authority.
+     *
+     * @param roleId    The unique identifier for the role.
+     * @param authority The authority (name) of the role.
+     */
+    public Role(long roleId, String authority) {
         this.id = roleId;
         this.authority = authority;
     }
